added bounded functions in python code
