+++ pydp/algorithms/laplacian/__init__.py
@@ -1,3 +1,4 @@
 from .count import Count
+from .boundedalgorithms import BoundedMean, BoundedSum, BoundedStandardDeviation, BoundedVariance
 
-__all__ = ["Count"]
+__all__ = ["Count", "BoundedMean", "BoundedSum", "BoundedStandardDeviation", "BoundedVariance"]
